+++ internal/app/sourcing.go
@@ -1,0 +1,5 @@
+package app
+
+func (receiver *Application) source() {
+	receiver.telegram.ForwardTo(receiver.routeChan)
+}
